RF: moved lines_to_vtk_polydata to utils.py
